# App/openapi.yaml
openapi: 3.0.3
info:
  title: Voenkom API
  version: 1.0.0
servers:
  - url: http://localhost:3001
    description: Local dev
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      properties:
        id:
          type: integer
        login:
          type: string
        name:
          type: string
        role:
          type: string
        email:
          type: string
...
paths:
...
  /api/users:
    get:
      security:
        - BearerAuth: []
      summary: List users
      responses:
        '200':
          description: Users array
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
    post:
      security:
        - BearerAuth: []
      summary: Create user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                login:
                  type: string
                password:
                  type: string
                name:
                  type: string
                role:
                  type: string
                email:
                  type: string
      responses:
        '200':
          description: Created user id
  /api/users/{id}:
    delete:
      security:
        - BearerAuth: []
      summary: Delete user (non-commissar)
...
  /api/personal-files/{id}:
    get:
      security:
        - BearerAuth: []
      summary: Get personal file by id
    put:
      security:
        - BearerAuth: []
      summary: Update personal file
  /api/summons:
    get:
      security:
        - BearerAuth: []
      summary: List summons
    post:
      security:
        - BearerAuth: []
      summary: Create summons
  /api/summons/{id}:
    put:
      security:
        - BearerAuth: []
      summary: Update summons
    delete:
      security:
        - BearerAuth: []
      summary: Delete summons
...
  /api/calendar:
    get:
      security:
        - BearerAuth: []
      summary: List calendar events
    post:
      security:
        - BearerAuth: []
      summary: Create calendar event
  /api/statistics:
    get:
      security:
        - BearerAuth: []
      summary: Statistics
  /api/audit:
    get:
      security:
        - BearerAuth: []
      summary: Audit log